added event handlers for stop and disconnect
I added the event handlers, and a dict keyed by a tuple of sid and port which regulates when the daemons should stop. I can't tell how much this is working because I haven't dove deeply into the front-end javascript, but some of the threads at least close so this seems like an improvement.